localentries: Move errno related functions to errno file
