--- internal/users/localentries/errno.go
@@ -1,5 +1,0 @@
-package localentries
-
-import "sync"
-
-var errnoMutex sync.Mutex
